dbReset file refactored to work
