--- backend/bin/resetdb.js
@@ -1,44 +1,0 @@
-require('dotenv').config();
-const fs = require('fs');
-const db = require('../src/db/connection');
-const chalk = require('chalk');
-
-
-// Loads the schema files from db/schema
-const runSchemaFiles = async () => {
-  console.log(chalk.cyan(`-> Loading Schema Files ...`));
-  const schemaFilenames = fs.readdirSync('./db/schema');
-
-  for (const fn of schemaFilenames) {
-    const sql = fs.readFileSync(`./db/schema/${fn}`, 'utf8');
-    console.log(`\t-> Running ${chalk.green(fn)}`);
-    await db.query(sql);
-  }
-};
-
-const runSeedFiles = async () => {
-  console.log(chalk.cyan(`-> Loading Seeds ...`));
-  const schemaFilenames = fs.readdirSync('./db/seeds');
-
-  for (const fn of schemaFilenames) {
-    const sql = fs.readFileSync(`./db/seeds/${fn}`, 'utf8');
-    console.log(`\t-> Running ${chalk.green(fn)}`);
-    await db.query(sql);
-  }
-};
-
-const runResetDB = async () => {
-  try {
-    process.env.DB_HOST &&
-      console.log(`-> Connecting to PG on ${process.env.DB_HOST} as ${process.env.DB_USER}...`);
-
-    await runSchemaFiles();
-    await runSeedFiles();
-    process.exit();
-  } catch (err) {
-    console.error(chalk.red(`Failed due to error: ${err}`));
-    process.exit();
-  }
-};
-
-runResetDB();
