+++ backend/src/db/schema/01_users.sql
@@ -1,6 +1,5 @@
 DROP TABLE IF EXISTS users CASCADE;
 
-
 CREATE TABLE users (
   id SERIAL PRIMARY KEY NOT NULL,
   username VARCHAR(255) NOT NULL,
