+++ backend/src/db/schema/02_workouts.sql
@@ -4,6 +4,6 @@ CREATE TABLE workouts (
   id SERIAL PRIMARY KEY NOT NULL,
   title VARCHAR(255) NOT NULL,
   is_complete BOOLEAN NOT NULL,
-  date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
-  user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
+  date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  user_id INTEGER REFERENCES users(id) ON DELETE CASCADE
 );
