+++ backend/src/db/schema/03_exercises.sql
@@ -5,5 +5,5 @@ CREATE TABLE exercises (
   title VARCHAR(255) NOT NULL,
   set_number INT NOT NULL,
   rep_number INT NOT NULL,
-  weight_number INT NOT NULL,
+  weight_number INT NOT NULL
 );
